캐시관리 업데이트

+++ process/@getCaches.php
@@ -8,35 +8,49 @@
  * @author Arzz (arzz@arzz.com)
  * @license MIT License
  * @version 7.1.2
- * @modified 2021. 5. 28.
+ * @modified 2021. 6. 3.
  */
 if (defined('__MINITALK__') == false) exit;
 
 $lists = array();
-
-foreach (GetDirectoryItems($this->getAttachmentPath().'/temp','file') as $file) {
+foreach (GetDirectoryItems($this->getAttachmentPath().'/cache','file') as $file) {
 	if (preg_match('/\.cache$/',$file) == true) {
 		$temp = explode('.',preg_replace('/\.cache$/','',basename($file)));
 		
-		$item = new stdClass();
-		$item->name = basename($file);
-		$item->type = array_pop($temp);
-		
-		$item->checksum = array_pop($temp);
-		$item->language = array_shift($temp);
-		$item->templet = array_shift($temp);
-		
-		if (count($temp) == 0 || strlen($item->checksum) != 6) {
-			unlink($file);
-			continue;
+		if ($temp[0] == 'core') {
+			$item = new stdClass();
+			$item->name = basename($file);
+			$item->type = 'json';
+			$item->checksum = null;
+			$item->language = '';
+			$item->templet = '';
+			$item->capacity = 'client';
+			$item->size = filesize($file);
+			$item->create_date = filemtime($file);
+			$item->path = $file;
+			
+			$lists[] = $item;
+		} else {
+			$item = new stdClass();
+			$item->name = basename($file);
+			$item->type = array_pop($temp);
+			
+			$item->checksum = array_pop($temp);
+			$item->language = array_shift($temp);
+			$item->templet = array_shift($temp);
+			
+			if (count($temp) == 0 || strlen($item->checksum) != 6) {
+				unlink($file);
+				continue;
+			}
+			
+			$item->capacity = implode('.',$temp);
+			$item->size = filesize($file);
+			$item->create_date = filemtime($file);
+			$item->path = $file;
+			
+			$lists[] = $item;
 		}
-		
-		$item->capacity = implode('.',$temp);
-		$item->size = filesize($file);
-		$item->create_date = filectime($file);
-		$item->path = $file;
-		
-		$lists[] = $item;
 	}
 }
